fix: LC-153 rename archivingActions file
